+++ src/IIS/Manager/Types/WebsiteManager.cs
@@ -17,7 +17,7 @@ namespace Cake.IIS
     /// </summary>
     public class WebsiteManager : BaseSiteManager
     {
-        #region Constructor (1)
+        #region Constructors
         /// <summary>
         /// Initializes a new instance of the <see cref="WebsiteManager" /> class.
         /// </summary>
@@ -34,7 +34,7 @@ namespace Cake.IIS
 
 
 
-        #region Methods (2)
+        #region Methods
         /// <summary>
         /// Creates a new instance of the <see cref="WebsiteManager" /> class.
         /// </summary>
@@ -65,18 +65,20 @@ namespace Cake.IIS
             if (!exists)
             {
                 _Server.CommitChanges();
+
                 // Settings which needs to be modified after the site is created.
                 var isModified = false;
                 if (settings.EnableDirectoryBrowsing)
                 {
                     var siteConfig = site.GetWebConfiguration();
-                    siteConfig.EnableDirectoryBrowsing(true);
+                    siteConfig.EnableDirectoryBrowsing();
                     isModified = true;
                 }
                 if (isModified)
                 {
                     _Server.CommitChanges();
                 }
+
                 _Log.Information("Web Site '{0}' created.", settings.Name);
             }
         }